requirements.txt
yfinance~=0.2.40
pandas~=2.2.1
numpy~=1.26.4
pyarrow~=15.0.2
boto3==1.34.151
s3fs>=2023.12.0
awswrangler==3.9.1
python-dotenv
psycopg2-binary
sqlalchemy